requirements.txt
flask
dnspython
flask-cors
